Refactor AI-enhanced messages into Messages component
Removed the AIEnhancedMessages component and moved its logic, data structures, and UI into the Messages component. This consolidates message triage, filtering, and display features into a single component for maintainability and reuse.
